Show warning message after generating an SSH key (#389) (#398)
Signed-off-by: Igor Vinokur <ivinokur@redhat.com>

Show's a warning message that SSH key will be available in all containers (not only in theia-ide container) only after restarting the workspace. Workspace restart will mount the new SSH keys to all containers. SSH plugin generates a new SSH key then it is written to the /home/theia/.ssh directory of the theia-ide container, so the key is invisible from other containers. The aim of the message is to tell the user that after restarting the workspace the key will be mounted to all containers of the workspace.