Apply GitHub authentication plugin (#604)
Add a plugin that authenticates vscode GitHub PR plugin launched in Che-Theia
The plugin calls Che oAuth API service to get the GitHub token. Then it injects it to the user preferences. When the browser page is refreshed the vscode GitHub PR plugin fetches the token and cleans the token from the preferences file.