Support encrypted SSH private keys (#719)
When uploading an encrypted SSH private key, ask a password for the key and register the key by executing ssh-add command. On the SSH plugin start iterate through the registered ssh keys and if a private key is encrypted read the passphrase from the key and register it with the help of ssh-add.